[RAPTOR-6439] handle defining bool args param through env var
