+++ custom_model_runner/requirements.txt
@@ -1,5 +1,7 @@
 argcomplete==1.11.1
 datarobot==2.24.0
+# trafaret version pinning is defined by `datarobot`
+trafaret!=1.1.0,<2.0,>=0.7
 docker>=4.2.2<5.0.0
 flask
 jinja2
